requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
firebase-admin==6.2.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
python-decouple==3.8
django-cors-headers==4.3.1
django-extensions==3.2.3
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
cryptography==41.0.7
phonenumbers==8.13.25
whatsapp-business==0.4.0
sentry-sdk==1.38.0
gunicorn==21.2.0
